made gpu use packing

+++ src/gpu.rs
@@ -197,10 +197,10 @@ impl<'gpu> GPUWork<'gpu> {
     }
 
     #[inline(always)]
-    pub(crate) fn calculate_matrix_mul(&mut self, matrix: &[[u16; 64]; 64]) {
+    pub(crate) fn calculate_matrix_mul(&mut self, matrix: &[[u8; 64]; 64]) {
         let func = &self.matrix_mul_kernel.func;
         let stream = &self.stream;
-        let mut matrix_gpu = self._module.get_global::<[[u16; 64]; 64]>(&CString::new("matrix").unwrap()).unwrap();
+        let mut matrix_gpu = self._module.get_global::<[[u8; 64]; 64]>(&CString::new("matrix").unwrap()).unwrap();
         matrix_gpu.copy_from(matrix);
         unsafe {
             launch!(
